FREEI-3352: if /AMPUSER/ext/accountcode is set to u+userId, set it as the accountcode when a conference room is called
This should be safe to merge whenever, since it doesn't rely on any
other code, and shouldn't change call behavior.  If a /AMPUSER has an accountcode set to the users
ID(FREEI-3350), set that as the accountcode when the conference room is
called.  The realtimeapi conference app looks for this to function
properly.

Please merge to 14 and up.